+++ app.py
@@ -2,9 +2,8 @@ from flask import Flask, render_template, session
 from view.create_account import create_account
 from view.login import login
 from view.group import group
-from view.add_job import add_job
+from view.job import job
 from view.edit_account import edit_account
-from view.edit_job import edit_job
 
 
 app = Flask(__name__)
@@ -25,15 +24,12 @@ def index():
 app.register_blueprint(create_account)
 # ログインページ
 app.register_blueprint(login)
-# バイト先情報登録画面
-app.register_blueprint(add_job)
+# バイト先ページ
+app.register_blueprint(job)
 # グループページ
 app.register_blueprint(group)
 # アカウント編集ページ
 app.register_blueprint(edit_account)
-# バイト先情報編集画面
-app.register_blueprint(edit_job)
-
 
 
 if __name__ == "__main__":
