+++ templates/add_job.html
@@ -19,7 +19,7 @@
 {% block main %}
 <main>
     <p id="error">{% if error %}{{error}}{% endif %}</p>
-    <form action="{{url_for('add_job.confirm_job_page')}}" method="post">
+    <form action="{{url_for('job.confirm_add_job_page')}}" method="post">
         <ul>
             <li>
                 <div class="input_text">
