+++ templates/confirm_edit_job.html
@@ -9,7 +9,7 @@
 {% endblock %}
 
 {% block back_page %}
-<a href="{{url_for('add_job.add_job_page')}}">＜</a>
+<a href="{{url_for('job.add_job_page')}}">＜</a>
 {% endblock %}
 
 {% block header %}
@@ -30,7 +30,7 @@
             </div>
         </dd>
     </dl>
-    <div><a href="{{url_for('edit_job.update_job')}}">保存する</a></div>
+    <div><a href="{{url_for('job.update_job')}}">保存する</a></div>
 </main>
 {% endblock %}
